requirements.txt
# Python3.5.2 or above is recommended
openpyxl==2.4.9
pandas==0.21.0
chardet==3.0.4
xlrd==1.1.0
selenium==3.8.1
pyodbc==4.0.24
PyMySQL==0.8.0
Jinja2==2.10
xlwt==1.3.0
psutil==5.6.0
PyAutoIt==0.4
Pillow==6.1.0
pywin32==224